Extract RichText so both page shells can share it
The /features pages render the same `**bold**` / `[label](/href)` copy the
/tools pages do, and that parser was a private function inside ToolPageShell.
Copying it would have meant two parsers that drift the moment one of them
learns a new token.

Pure move: identical implementation, ToolPageShell now imports it.
